Moved mockRecipe
