+++ ff.web/tsconfig.json
@@ -32,6 +32,11 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src", "src/custom.d.ts", "../shared-components/src/QuestLineOnboardingHero.tsx", ".storybook/"],
+  "include": [
+    "src",
+    "src/custom.d.ts",
+    "../shared-components/src/QuestLineOnboardingHero.tsx",
+    ".storybook/",
+  ],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
